requirements.txt
gitpython>=3.1.0
commitizen>=2.19.0
pyyaml>=5.4.1
click>=8.0.0
pytest>=7.0.0
